pyproject.toml
[project]
name = "mcp-timeplus"
version = "0.1.4"
description = "An MCP server for Timeplus."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE"]
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.3.0",
    "mcp[cli]>=1.3.0",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.34.0",
    "timeplus-connect>=0.8.15",
    "confluent-kafka",
    "pip-system-certs>=4.0",
]

[project.scripts]
mcp-timeplus = "mcp_timeplus.main:main"

[project.urls]
Home = "https://github.com/jovezhong/mcp-timeplus"

[project.optional-dependencies]
dev = ["ruff", "pytest"]

[tool.hatch.build.targets.wheel]
packages = ["mcp_timeplus"]

[tool.ruff]
line-length = 100

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
